Created a Course and Review POJO
Added dll script init.sql to create database table and columns

Added CourseDAO and ReviewDAO interfaces with methods to add, and query Review and courses.

Added Sql2CourseDAO that implemented the CourseDAO interface
with findAll() to return a list of Courses, add() to add a single course

Added a test class

Added a test method to check that the add() method of the Sql2CourseDAO method changes the id value of the course we just added.

Added a test method to check that the findAll method returns a list containing the course we added.